+++ pkg/streamingpromql/operators/selectors/range_vector_selector.go
@@ -7,6 +7,7 @@ package selectors
 
 import (
 	"context"
+	"errors"
 	"fmt"
 
 	"github.com/prometheus/prometheus/model/value"
@@ -22,22 +23,33 @@ type RangeVectorSelector struct {
 	Selector *Selector
 	Stats    *types.QueryStats
 
-	rangeMilliseconds int64
-	chunkIterator     chunkenc.Iterator
-	nextStepT         int64
-	floats            *types.FPointRingBuffer
-	histograms        *types.HPointRingBuffer
-	stepData          *types.RangeVectorStepData // Retain the last step data instance we used to avoid allocating it for every step.
+	rangeMilliseconds   int64
+	chunkIterator       chunkenc.Iterator
+	nextStepT           int64
+	floats              *types.FPointRingBuffer
+	extendedRangeFloats *types.FPointRingBuffer // A buffer we use to create views for smoothed/anchored extended ranges which have added/modified points from the original floats buffer
+	histograms          *types.HPointRingBuffer
+	stepData            *types.RangeVectorStepData // Retain the last step data instance we used to avoid allocating it for every step.
+	reusableView        *types.FPointRingBufferView
+
+	memoryConsumptionTracker *limiter.MemoryConsumptionTracker
+	anchored                 bool // The anchored modifier has been used for this range query
+	smoothed                 bool // The smoothed modifier has been used for this range query
 }
 
 var _ types.RangeVectorOperator = &RangeVectorSelector{}
 
-func NewRangeVectorSelector(selector *Selector, memoryConsumptionTracker *limiter.MemoryConsumptionTracker) *RangeVectorSelector {
+func NewRangeVectorSelector(selector *Selector, memoryConsumptionTracker *limiter.MemoryConsumptionTracker, anchored bool, smoothed bool) *RangeVectorSelector {
+
 	return &RangeVectorSelector{
-		Selector:   selector,
-		floats:     types.NewFPointRingBuffer(memoryConsumptionTracker),
-		histograms: types.NewHPointRingBuffer(memoryConsumptionTracker),
-		stepData:   &types.RangeVectorStepData{},
+		Selector:                 selector,
+		floats:                   types.NewFPointRingBuffer(memoryConsumptionTracker),
+		histograms:               types.NewHPointRingBuffer(memoryConsumptionTracker),
+		extendedRangeFloats:      types.NewFPointRingBuffer(memoryConsumptionTracker),
+		stepData:                 &types.RangeVectorStepData{Anchored: anchored, Smoothed: smoothed}, // Include the smoothed/anchored context to the step data as functions such as rate/increase require this
+		anchored:                 anchored,
+		smoothed:                 smoothed,
+		memoryConsumptionTracker: memoryConsumptionTracker,
 	}
 }
 
@@ -62,6 +74,7 @@ func (m *RangeVectorSelector) NextSeries(ctx context.Context) error {
 	m.nextStepT = m.Selector.TimeRange.StartT
 	m.floats.Reset()
 	m.histograms.Reset()
+	m.extendedRangeFloats.Reset()
 	return nil
 }
 
@@ -82,17 +95,70 @@ func (m *RangeVectorSelector) NextStepSamples(ctx context.Context) (*types.Range
 	// Apply offset after adjusting for timestamp from @ modifier.
 	rangeEnd = rangeEnd - m.Selector.Offset
 	rangeStart := rangeEnd - m.rangeMilliseconds
+
+	// Take a copy of the original range - the smoothed/anchored modifiers will change these
+	originalRangeStart := rangeStart
+	originalRangeEnd := rangeEnd
+
+	// When the smoothed/anchored modifiers are used, the selector (fillBuffer) will return a wider range of points.
+	// The range boundaries are modified accordingly so that we do not discard these extended points.
+	if m.anchored {
+		rangeStart -= m.Selector.LookbackDelta.Milliseconds()
+	} else if m.smoothed {
+		rangeStart -= m.Selector.LookbackDelta.Milliseconds()
+		rangeEnd += m.Selector.LookbackDelta.Milliseconds()
+	}
+
 	m.floats.DiscardPointsAtOrBefore(rangeStart)
 	m.histograms.DiscardPointsAtOrBefore(rangeStart)
+	m.stepData.SmoothedBasisForHeadPoint = nil
+	m.stepData.SmoothedBasisForTailPoint = nil
 
+	// Fill the buffer with an extended range of points (smoothed/anchored) - these will be filtered out in the extendRangeVectorPoints() below
 	if err := m.fillBuffer(m.floats, m.histograms, rangeStart, rangeEnd); err != nil {
 		return nil, err
 	}
 
-	m.stepData.Floats = m.floats.ViewUntilSearchingBackwards(rangeEnd, m.stepData.Floats)
+	if m.anchored || m.smoothed {
+
+		// Release the temporary ring buffer and initialise it off our given buff.
+		// The ring buffer will release the buff back to the slice pool.
+		m.extendedRangeFloats.Release()
+
+		// Histograms are not supported for these modified range queries
+		if m.histograms.ViewUntilSearchingForwards(rangeEnd, nil).Count() > 0 {
+			return nil, errors.New("smoothed and anchored modifiers do not work with native histograms")
+		}
+
+		// Note the extended range end is used since smoothed will have extended this
+		m.reusableView = m.floats.ViewUntilSearchingForwards(rangeEnd, m.reusableView)
+
+		// buff is a new slice of points which includes points for the range boundaries
+		// smoothedHead/Tail are special cases of the boundary points which will be used by any rate/increase function which consumes this vector.
+		buff, smoothedHead, smoothedTail, err := extendRangeVectorPoints(m.reusableView, originalRangeStart, originalRangeEnd, m.smoothed, m.memoryConsumptionTracker)
+		if err != nil {
+			return nil, err
+		}
+
+		if buff != nil {
+			err := m.extendedRangeFloats.Use(buff)
+			if err != nil {
+				return nil, err
+			}
+		}
+
+		// Store the smoothed points in the range step data result so that consumers of this data can reference these values
+		// without having to re-calculate off the original points. Re-use the view
+		m.stepData.Floats = m.extendedRangeFloats.ViewAll(m.stepData.Floats)
+		m.stepData.SmoothedBasisForHeadPoint = smoothedHead
+		m.stepData.SmoothedBasisForTailPoint = smoothedTail
+	} else {
+		m.stepData.Floats = m.floats.ViewUntilSearchingBackwards(rangeEnd, m.stepData.Floats)
+	}
+
 	m.stepData.Histograms = m.histograms.ViewUntilSearchingBackwards(rangeEnd, m.stepData.Histograms)
-	m.stepData.RangeStart = rangeStart
-	m.stepData.RangeEnd = rangeEnd
+	m.stepData.RangeStart = originalRangeStart // important to return the original range start so that functions like rate() can determine the range duration regardless of smoothed / anchored
+	m.stepData.RangeEnd = originalRangeEnd
 
 	m.Stats.IncrementSamplesAtTimestamp(m.stepData.StepT, int64(m.stepData.Floats.Count())+m.stepData.Histograms.EquivalentFloatSampleCount())
 
@@ -166,6 +232,7 @@ func (m *RangeVectorSelector) Finalize(ctx context.Context) error {
 func (m *RangeVectorSelector) Close() {
 	m.Selector.Close()
 	m.floats.Close()
+	m.extendedRangeFloats.Close()
 	m.histograms.Close()
 	m.chunkIterator = nil
 }